wallet controller

+++ database/migrations/2025_08_02_000007_create_wallets_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
-            $table->integer('total_points')->default(0);
+            $table->decimal('total_points', 18, 2)->default(0);
 
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
